+++ lib/cache_x.dart
@@ -4,3 +4,4 @@ import 'package:cache_x/src/cacheXWorker.dart';
 import 'package:flutter/material.dart';
 
 part 'src/cacheXCore.dart';
+part 'src/exception.dart';