+++ lib/src/exception.dart
@@ -1,3 +1,5 @@
+part of cache_x;
+
 /// Exception Related to ChacheX Storage
 class StorageException implements Exception {
   final dynamic message;
@@ -17,5 +19,16 @@ class EncryptionException implements Exception {
     Object message = this.message;
     if (message == null) return "EncryptionException";
     return "EncryptionException: $message";
+  }
+}
+
+/// Exception related to CacheX Core
+class CacheXException implements Exception {
+  final dynamic message;
+  CacheXException([this.message]);
+  String toString() {
+    Object message = this.message;
+    if (message == null) return "EncryptionException";
+    return "CacheXException: $message";
   }
 }